+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation_objc/include/video_player_avfoundation_objc/FVPVideoPlayer.h
@@ -29,6 +29,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property(nonatomic, readonly) BOOL disposed;
 /// Indicates whether the video player is set to loop.
 @property(nonatomic) BOOL isLooping;
+/// The preferred audio language code
+@property(nonatomic, copy, nullable) NSString *preferredAudioLanguage;
 /// The current playback position of the video, in milliseconds.
 @property(nonatomic, readonly) int64_t position;
 /// The event listener to report video events to.