+++ packages/video_player/video_player_avfoundation/lib/src/video_player_plugin_messages.g.dart
@@ -135,14 +135,16 @@ class PlatformVideoViewCreationParams {
 }
 
 class CreationOptions {
-  CreationOptions({required this.uri, required this.httpHeaders});
+  CreationOptions({required this.uri, required this.httpHeaders, this.preferredAudioLanguage});
 
   String uri;
 
   Map<String, String> httpHeaders;
 
+  String? preferredAudioLanguage;
+
   List<Object?> _toList() {
-    return <Object?>[uri, httpHeaders];
+    return <Object?>[uri, httpHeaders, preferredAudioLanguage];
   }
 
   Object encode() {
@@ -154,6 +156,7 @@ class CreationOptions {
     return CreationOptions(
       uri: result[0]! as String,
       httpHeaders: (result[1]! as Map<Object?, Object?>).cast<String, String>(),
+      preferredAudioLanguage: result[2] as String?,
     );
   }
 
@@ -166,7 +169,9 @@ class CreationOptions {
     if (identical(this, other)) {
       return true;
     }
-    return _deepEquals(uri, other.uri) && _deepEquals(httpHeaders, other.httpHeaders);
+    return _deepEquals(uri, other.uri) &&
+        _deepEquals(httpHeaders, other.httpHeaders) &&
+        _deepEquals(preferredAudioLanguage, other.preferredAudioLanguage);
   }
 
   @override