+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.10.0
+
+* Adds support for setting a preferred audio language during player creation via `VideoPlayerOptions preferredAudioLanguage`.
+
 ## 6.9.0
 
 * Adds `backBufferDurationMs` to `VideoPlayerOptions` to support configuring the back buffer duration.